# src/api/contracts/school-openapi.yaml
swagger: "2.0"
info:
  title: School API
  description: School API
  version: 1.0.5
  contact:
    email: orscarsilla@outlook.com
...
consumes:
  - application/json
produces:
  - application/json
paths:
  /users:
    post:
      tags:
        - "Users"
      summary: "Create user"
      description: "Create a new user (Register)"
      operationId: "createUser"
      parameters:
        - $ref: "#/parameters/userBodyParam"
      responses:
        '201':
          description: Created successfully.
          schema:
            $ref: "#/definitions/GeneratedIdResponse"
        '400':
          description: Bad request.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description: Unauthorized.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: Not found.
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      tags:
        - "Users"
      summary: "Get all users"
      description: "Get all users"
      operationId: "getUsers"
      responses:
        '200':
          description: Ok.
          schema:
            $ref: "#/definitions/UsersResponse"
        '400':
          description: Bad request.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description: Unauthorized.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: Not found.
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users/{id}:
    get:
      tags:
        - "Users"
      summary: "Get user"
      description: "Get one user by _id"
      operationId: "getUser"
      parameters:
        - $ref: "#/parameters/idParam"
      responses:
        '200':
          description: Ok.
          schema:
            $ref: "#/definitions/UserResponse"
        '400':
          description: Bad request.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description: Unauthorized.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: Not found.
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      tags:
        - "Users"
      summary: "Delete user"
      description: "Delete user by id"
      operationId: "deleteUser"
      parameters:
        - $ref: "#/parameters/idParam"
      responses:
        '204':
          description: No Content.
        '400':
          description: Bad request.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description: Unauthorized.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: Not found.
          schema:
            $ref: "#/definitions/ErrorResponse"
  /video:
    post:
      tags:
        - "Video"
      summary: "Create video"
      description: "Create video"
      operationId: "createVideo"
      parameters:
        - $ref: "#/parameters/videoBodyParam"
      responses:
        '201':
          description: Created successfully.
          schema:
            $ref: "#/definitions/GeneratedIdResponse"
        '400':
          description: Bad request.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description: Unauthorized.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: Not found.
          schema:
            $ref: "#/definitions/ErrorResponse"
  /video/{id}:
    patch:
      tags:
        - "Video"
      summary: "Modify an existing video"
      description: "Modify 'src' from video."
      operationId: "modifyVideo"
      parameters:
        - $ref: "#/parameters/idParam"
        - $ref: "#/parameters/videoSourceBodyParam"
      responses:
        '200':
          description: OK.
        '400':
          description: Bad request.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description: Unauthorized.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: Not found.
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      tags:
        - "Video"
      summary: "Delete video"
      description: "Delete video by id"
      operationId: "deleteVideo"
      parameters:
        - $ref: "#/parameters/idParam"
      responses:
        '204':
          description: No content.
        '400':
          description: Bad request.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description: Unauthorized.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: Not found.
          schema:
            $ref: "#/definitions/ErrorResponse"
  /video/{ref}:
    get:
      tags:
        - "Video"
      summary: "Get an existing video"
      description: "Get existing video by ObjectId."
      operationId: "getVideo"
      parameters:
        - $ref: "#/parameters/idParam"
      responses:
        '200':
          description: OK.
          schema:
            $ref: '#/definitions/VideoResponse'
        '400':
          description: Bad request.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description: Unauthorized.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: Not found.
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  VideoResponse:
    type: object
    required:
      - "_id"
      - "ref"
      - "src"
    properties:
      _id:
        type: string
        format: string
        example: "507f191e810c19729de860ea"
      ref:
        type: string
        format: string
        example: home
      src:
        type: string
        format: string
        example: <iframe width="560" height="315" src="https://www.youtube.com/embed/d6q8T2jfjm0" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
  ErrorResponse:
    type: object
    required:
      - "code"
      - "description"
    properties:
      code:
        type: integer
        format: int32
      description:
        type: string
        format: string
  VideoBody:
    type: object
    required:
      - "ref"
      - "src"
    properties:
      ref:
        type: string
        format: string
        example: home
      src:
        type: string
        format: string
        example: <iframe width="560" height="315" src="https://www.youtube.com/embed/d6q8T2jfjm0" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
  VideoSourceBody:
    type: object
    required:
      - "src"
    properties:
      src:
        type: string
        format: string
        example: <iframe width="560" height="315" src="https://www.youtube.com/embed/d6q8T2jfjm0" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
...
  UserBody:
    type: object
    required:
      - "username"
      - "password"
      - "name"
      - "email"
    properties:
      username:
        type: string
        format: string
        example: "osilla"
      password:
        type: string
        format: string
        example: "1234"
      name:
        type: string
        format: string
        example: "Don Quijote"
      email:
        type: string
        format: string
        example: "oscarsilla@outlook.com"
      surnames:
        type: string
        format: string
        example: "De la mancha"
      avatar:
        type: string
        format: string
        example: "avatar.png"
...
  UsersResponse:
    type: array
    items:
      $ref: "#/definitions/UserResponse"
  UserResponse:
    type: object
    properties:
      _id:
        type: string
        format: string
        example: 507f191e810c19729de860ea
      username:
        type: string
        format: string
        example: "osilla"
      password:
        type: string
        format: string
        example: "1234"
      name:
        type: string
        format: string
        example: "Don Quijote"
      email:
        type: string
        format: string
        example: "oscarsilla@outlook.com"
      surnames:
        type: string
        format: string
        example: "De la mancha"
      avatar:
        type: string
        format: string
        example: "avatar.png"
  GeneratedIdResponse:
    type: object
    properties:
      generatedId:
        type: string
        format: string
        example: "63b04cf26cff5203de1659ca"
parameters:
  idParam:
    name: "_id"
    description: "ObjectId param."
    in: path
    required: true
    type: string
    format: string
...
  videoBodyParam:
    name: "body"
    description: "Video body"
    in: body
    required: true
    schema:
      $ref: "#/definitions/VideoBody"
  videoSourceBodyParam:
    name: "body"
    description: "Video source body param."
    in: body
    required: true
    schema:
      $ref: "#/definitions/VideoSourceBody"
...
  userBodyParam:
    name: "body"
    description: "User body param"
    in: body
    required: true
    schema:
      $ref: "#/definitions/UserBody"
...